+++ changelog.md
@@ -13,6 +13,7 @@
 ## Improvements
 
 * Allow for `Node` and `Relationship` inputs to be given as `camelCase` and `SCREAMING_SNAKE_CASE` (in addition to `snake_case`).
+* Convert non-json serializable properties to strings in the `render` method, instead of raising an error.
 
 
 ## Other changes